Add engine class and update dashboard UI
The engine class has been introduced with a basic test, allowing the creation of engine objects. The dashboard interface received significant updates allowing it to better display and interact with these engine objects. The UI components are now built more dynamically, adapting better to the content, and there's now an indication of when an object is of the 'engine' type. Minor bug fixes and code refactor also took place.