Update mixin styles and improve UI design
Enhancements were implemented in the color scheme, scrollbar, and shadow styling, mainly within the 'styles.scss' and 'data-dashboard.component.scss' files. A new dark primary color was introduced, and a mix-in for an inset shadow was added to boost UI aesthetic. The custom scrollbar design was also made consistent to enhance user experience and interaction.