Refactor AttributeMethod class and update ToolsService
The AttributeMethod class has been refactored for better organization and readability. In the constructor, the assignment of values to class properties was moved up for visibility. Some methods and variables in the AttributeMethod class were rearranged and comments were enhanced for better context. In ToolsService, static properties and factory methods were moved to the top for better accessibility. Corrections were also made to styles in the first-tool.component.ts file.