Added GroupCount
This commit implements a new GroupCount method in the tools service, which groups and counts properties. Additional modifications include adjusting types in the GetArrayOfProperty method for better clarity and flexibility. A new Initializer method was also introduced in the FirstToolComponent, which integrates the use of the GetArrayOfProperty and GroupCount methods into the tool's initialization process. This is reflected in the count of names visible in the user interface.